Created LiveChatIncTypingSpeedTestDestroyed.py
The script opens Live Chat Inc Typing Speed Test and completes its typing test all by itself destroying it in 50 seconds. The script is able to score a speed of 507 WPM (2259 CPM) with an accuracy of 100%. The typing test goes out of words and the script is idle for the last 10-12 seconds of the total 60 seconds because of the scarcity of words. LOL.